feat: allow decimal (0.5) quantities in audit counts
- audit_mismatches.expected/actual changed from integer to numeric(10,2)
- Count inputs now accept decimals (step=0.5, min=0)
- Removed Math.round() and digit-only restriction added in prev commit

+++ supabase/schema.sql
@@ -141,8 +141,8 @@ create table public.audit_mismatches (
   item_id uuid references public.inventory(id) on delete set null,
   name text not null,
   sku text not null,
-  expected integer not null,
-  actual integer not null,
+  expected numeric(10,2) not null,
+  actual numeric(10,2) not null,
   remark text
 );
